Finish fixing compatability!
Was bored and clicked on the wikipedia link on the readme to chip8. I then looked at op codes and for some reason decided to switch theorder of setting the vF register and the x/y register. This then just solved half the issues. The rest were solved using similar methods